+++ pypeit/tests/test_load_images.py
@@ -148,6 +148,15 @@ def test_load_gmos():
         pytest.fail('Gemini GMOS test data section failed: {0}'.format(ifile))
 
 @dev_suite_required
+def test_load_osiris():
+    ifile = os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA/gtc_osiris/R2500R',
+                         '0002851159-20210217-OSIRIS-OsirisBias.fits')
+    try:
+        data_img = grab_img('gtc_osiris', ifile)
+    except:
+        pytest.fail('GTC OSIRIS test data section failed: {0}'.format(ifile))
+
+@dev_suite_required
 def test_load_bok():
     ifile = os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA/bok_bc/600',
                          'g0005.fits')
@@ -178,7 +187,7 @@ def test_load_hires():
         data_img = proc.raw_images[0][proc.datasec[0][0]]
     except:
         pytest.fail('Keck HIRES test data section failed: {0}'.format(files))
-        
+
 @dev_suite_required
 def test_load_isis():
     files = os.path.join(os.getenv('PYPEIT_DEV'), 'RAW_DATA', 'wht_isis_blue', 'long_R300B_d5300',
@@ -191,4 +200,3 @@ def test_load_isis():
     except:
         pytest.fail('WHT ISIS test data section failed.')
 '''
-